Added Application Library with working with Business Layer(Logic). Added Domain Library where we will save our models.
Onion Architecture finished on 50%
