+++ Context/TodoContext.cs
@@ -1,0 +1,15 @@
+﻿using Microsoft.EntityFrameworkCore;
+using TodoApi.Models;
+
+namespace TodoApiDTO.Context
+{
+  public class TodoContext : DbContext
+  {
+    public TodoContext(DbContextOptions<TodoContext> options)
+        : base(options)
+    {
+    }
+
+    public DbSet<TodoItem> TodoItems { get; set; }
+  }
+}
