--- Models/TodoContext.cs
@@ -1,14 +1,0 @@
-﻿using Microsoft.EntityFrameworkCore;
-
-namespace TodoApi.Models
-{
-    public class TodoContext : DbContext
-    {
-        public TodoContext(DbContextOptions<TodoContext> options)
-            : base(options)
-        {
-        }
-
-        public DbSet<TodoItem> TodoItems { get; set; }
-    }
-}
