+++ Models/TodoItem.cs
@@ -1,8 +1,13 @@
-﻿namespace TodoApi.Models
+﻿using System.ComponentModel.DataAnnotations;
+using System.ComponentModel.DataAnnotations.Schema;
+
+namespace TodoApi.Models
 {
     #region snippet
     public class TodoItem
     {
+        [Key]
+        [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
         public long Id { get; set; }
         public string Name { get; set; }
         public bool IsComplete { get; set; }
