--- Models/TodoItemDTO.cs
@@ -1,11 +1,0 @@
-﻿namespace TodoApi.Models
-{
-    #region snippet
-    public class TodoItemDTO
-    {
-        public long Id { get; set; }
-        public string Name { get; set; }
-        public bool IsComplete { get; set; }
-    }
-    #endregion
-}
