+++ TodoApiDTO.sln
@@ -1,9 +1,13 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.30002.166
+# Visual Studio Version 17
+VisualStudioVersion = 17.0.32014.148
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TodoApiDTO", "TodoApiDTO.csproj", "{623124F9-F5BA-42DD-BC26-A1720774229C}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Application", "..\Application\Application.csproj", "{83351D74-DB4D-445B-B8B0-B173A250F3D0}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Domain", "..\Domain\Domain.csproj", "{17A9E309-C163-484D-8269-63BBE709A27D}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -15,6 +19,14 @@ Global
 		{623124F9-F5BA-42DD-BC26-A1720774229C}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{623124F9-F5BA-42DD-BC26-A1720774229C}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{623124F9-F5BA-42DD-BC26-A1720774229C}.Release|Any CPU.Build.0 = Release|Any CPU
+		{83351D74-DB4D-445B-B8B0-B173A250F3D0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{83351D74-DB4D-445B-B8B0-B173A250F3D0}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{83351D74-DB4D-445B-B8B0-B173A250F3D0}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{83351D74-DB4D-445B-B8B0-B173A250F3D0}.Release|Any CPU.Build.0 = Release|Any CPU
+		{17A9E309-C163-484D-8269-63BBE709A27D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{17A9E309-C163-484D-8269-63BBE709A27D}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{17A9E309-C163-484D-8269-63BBE709A27D}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{17A9E309-C163-484D-8269-63BBE709A27D}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
